separate training, validation, predict process
